[tools] perf tools: Improve thread_stack__process() for trace begin / end
Message-id: <b6576f62facbedfe584f6cc3c2f23a4190ddf59b.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261455
O-Subject: [RHEL8.1 BZ1709243 131/217] perf tools: Improve thread_stack__process() for trace begin / end
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit 2dcde4e152a3e319cc7e76c7c6b8548a3c72310d
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Sep 20 16:00:46 2018 +0300

description
===========
thread_stack__process() is used to create call paths for database
export.  Improve the handling of trace begin / end to allow for a trace
that ends in a call.

Previously, the Intel PT decoder would indicate begin / end by a branch
from / to zero. That hides useful information, in particular when a
trace ends with a call. Before remedying that, enhance the thread stack
so that it identifies the trace end by the flag instead of by ip == 0.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>