[tools] tools include: Adopt linux/bits.h
Message-id: <4a1bd1de4d95a6bec703ad73710409f761126066.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261483
O-Subject: [RHEL8.1 BZ1709243 134/217] tools include: Adopt linux/bits.h
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit ba4aa02b417f08a0bee5e7b8ed70cac788a7c854
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Sep 25 10:55:59 2018 -0300

description
===========
So that we reduce the difference of tools/include/linux/bitops.h to the
original kernel file, include/linux/bitops.h, trying to remove the need
to define BITS_PER_LONG, to avoid clashes with asm/bitsperlong.h.

And the things removed from tools/include/linux/bitops.h are really in
linux/bits.h, so that we can have a copy and then
tools/perf/check_headers.sh will tell us when new stuff gets added to
linux/bits.h so that we can check if it is useful and if any adjustment
needs to be done to the tools/{include,arch}/ copies.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>