[tools] perf evsel: Mark a evsel as disabled when asking the kernel do disable it
Message-id: <e87d259fdf3caef1eae9a34885b6be90107a301a.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261415
O-Subject: [RHEL8.1 BZ1709243 154/217] perf evsel: Mark a evsel as disabled when asking the kernel do disable it
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit b7e8452b860c299f342a012922bdd9ab8f2bb722
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Sat Oct 20 09:04:41 2018 -0300

description
===========
Because there may be more such events in the ring buffer that should be
discarded when an app decides to stop considering them.

At some point we'll do this with eBPF, this way we stop them at origin,
before they are placed in the ring buffer.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>