[tools] perf trace: Drop addr_location refcounts
Message-id: <4c12c9b83e36c63748670e3c3d840f1231233763.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261416
O-Subject: [RHEL8.1 BZ1709243 155/217] perf trace: Drop addr_location refcounts
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit 4291bf5cb93918232f88a3a70d8f70a72fbb6ab0
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Sat Oct 20 09:18:50 2018 -0300

description
===========
When we use machine__resolve() we grab a reference to
addr_location.thread (and in the future to other elements there) via
machine__findnew_thread(), so we must pair that with
addr_location__put(), else we'll never drop that thread when it exits
and no other remaining data structures have pointers to it. Fix it.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>