[tools] tools include uapi: Grab a copy of linux/fs.h
Message-id: <42c4ab1f35623ca64d9761e18ade06c36106e224.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261440
O-Subject: [RHEL8.1 BZ1709243 180/217] tools include uapi: Grab a copy of linux/fs.h
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit f443f38c5789ece6ebe59ae21c27bf861e61c4e2
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Oct 24 14:31:12 2018 -0300

description
===========
We'll use it to create tables for the 'flags' argument to the 'mount'
and 'umount' syscalls.

Add it to check_headers.sh so that when a new protocol gets added we get
a notification during the build process.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>