[tools] perf trace beauty: Allow syscalls to mask an argument before considering it
Message-id: <870868ba8df2caa8378af30784f6f9e5d34f579e.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261458
O-Subject: [RHEL8.1 BZ1709243 184/217] perf trace beauty: Allow syscalls to mask an argument before considering it
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit 496fd346b71ffa0ff35623e9ec79df0bad47bd66
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Oct 25 16:09:47 2018 -0300

description
===========
Take mount's 'flags' arg, to cope with this semantic, as defined in do_mount in fs/namespace.c:

  /*
   * Pre-0.97 versions of mount() didn't have a flags word.  When the
   * flags word was introduced its top half was required to have the
   * magic value 0xC0ED, and this remained so until 2.4.0-test9.
   * Therefore, if this magic number is present, it carries no
   * information and must be discarded.
   */

We need to mask this arg, and then see if it is zero, when we simply
don't print the arg name and value.

The next patch will use this for mount's 'flag' arg.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>