[tools] perf callchain: Honour the ordering of PERF_CONTEXT_{USER,KERNEL,etc}
Message-id: <edb453862c00dac3acfba3c83aec300dda690e3d.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261553
O-Subject: [RHEL8.1 BZ1709243 192/217] perf callchain: Honour the ordering of PERF_CONTEXT_{USER, KERNEL, etc}
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit e9024d519d892b38176cafd46f68a7cdddd77412
Author: David S. Miller <davem@davemloft.net>
Date: Tue Oct 30 12:12:26 2018 -0300

description
===========
When processing using 'perf report -g caller', which is the default, we
ended up reverting the callchain entries received from the kernel, but
simply reverting throws away the information that tells that from a
point onwards the addresses are for userspace, kernel, guest kernel,
guest user, hypervisor.

The idea is that if we are walking backwards, for each cluster of
non-cpumode entries we have to first scan backwards for the next one and
use that for the cluster.

This seems silly and more expensive than it needs to be but it is enough
for a initial fix.

The code here is really complicated because it is intimately intertwined
with the lbr and branch handling, as well as this callchain order,
further fixes will be needed to properly take into account the cpumode
in those cases.

Another problem with ORDER_CALLER is that the NULL "0" IP that is at the
end of most callchains shows up at the top of the histogram because
every callchain contains it and with ORDER_CALLER it is the first entry.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>