[tools] perf trace: Fix setting of augmented payload when using eBPF + raw_syscalls
Message-id: <42ef6420254de694268b8e2175882c5c0a048721.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261539
O-Subject: [RHEL8.1 BZ1709243 201/217] perf trace: Fix setting of augmented payload when using eBPF + raw_syscalls
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit cd26ea6d50a207ee37e0364ecc2d196d6c9671e8
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Sat Nov 3 08:19:56 2018 -0300

description
===========
For now with BPF raw_augmented we hook into raw_syscalls:sys_enter and
there we get all 6 syscall args plus the tracepoint common fields
(sizeof(long)) and the syscall_nr (another long). So we check if that is
the case and if so don't look after the sc->args_size, but always after
the full raw_syscalls:sys_enter payload, which is fixed.

We'll revisit this later to pass s->args_size to the BPF augmenter (now
tools/perf/examples/bpf/augmented_raw_syscalls.c, so that it copies only
what we need for each syscall, like what happens when we use
syscalls:sys_enter_NAME, so that we reduce the kernel/userspace traffic
to just what is needed for each syscall.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>