[tools] perf tools: Restore proper cwd on return from mnt namespace
Message-id: <f7b3d862ffe5ddcd39fcf3c4552637df738003f5.1559304074.git.mpetlan@redhat.com>
Patchwork-id: 261532
O-Subject: [RHEL8.1 BZ1709243 216/217] perf tools: Restore proper cwd on return from mnt namespace
Bugzilla: 1709243
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1709243
https://bugzilla.redhat.com/show_bug.cgi?id=1709243

upstream
========
commit b01c1f69c8660eaeab7d365cd570103c5c073a02
Author: Jiri Olsa <jolsa@kernel.org>
Date: Thu Nov 1 18:00:01 2018 +0100

description
===========
When reporting on 'record' server we try to retrieve/use the mnt
namespace of the profiled tasks. We use following API with cookie to
hold the return namespace, roughly:

  nsinfo__mountns_enter(struct nsinfo *nsi, struct nscookie *nc)
    setns(newns, 0);
  ...
  new ns related open..
  ...
  nsinfo__mountns_exit(struct nscookie *nc)
    setns(nc->oldns)

Once finished we setns to old namespace, which also sets the current
working directory (cwd) to "/", trashing the cwd we had.

This is mostly fine, because we use absolute paths almost everywhere,
but it screws up 'perf diff':

  # perf diff
  failed to open perf.data: No such file or directory  (try 'perf record' first)
  ...

Adding the current working directory to be part of the cookie and
restoring it in the nsinfo__mountns_exit call.

Fixes: 843ff37bb59e ("perf symbols: Find symbols in different mount namespace")
[ No need to check for NULL args for free(), use zfree() for struct members ]

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>