[scsi] scsi: megaraid_sas: Add support for DEVICE_LIST DCMD in driver
Message-id: <20190531134850.7445-31-thenzl@redhat.com>
Patchwork-id: 261601
O-Subject: [RHEL8.1 e-stor 30/39] scsi: megaraid_sas: Add support for DEVICE_LIST DCMD in driver
Bugzilla: 1668409
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This patch adds support for the new DEVICE_LIST DCMD.

Driver currently sends two separate DCMDs for getting the list of PDs and
LDs that are exposed to host.  The new DCMD provides a single interface to
get a list of both PDs and LDs that are exposed to the host.  Based on the
list of target IDs that are returned by this DCMD, driver will add the
devices (PD/LD) to SML.  Driver will check for FW support for this new DCMD
and based on the support will either send the new DCMD or will fall back to
the earlier method of sending two separate DCMDs for PD and LD list.

Signed-off-by: Kashyap Desai <kashyap.desai@broadcom.com>
Signed-off-by: Shivasharan S <shivasharan.srikanteshwara@broadcom.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit f6fe57310811780d55d79e30da51db98677f1a90)
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>