[nvme] nvme-multipath: split bios with the ns_head bio_set before submitting
Message-id: <20190605152531.5352-3-emilne@redhat.com>
Patchwork-id: 262788
O-Subject: [RHEL8.1 PATCH 2/7] nvme-multipath: split bios with the ns_head bio_set before submitting
Bugzilla: 1715135
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

If the bio is moved to a different queue via blk_steal_bios() and
the original queue is destroyed in nvme_remove_ns() we'll be ending
with a crash in bio_endio() as the mempool for the split bio bvecs
had already been destroyed.
So split the bio using the original queue (which will remain during the
lifetime of the bio) before sending it down to the underlying device.

Signed-off-by: Hannes Reinecke <hare@suse.com>
Reviewed-by: Ming Lei <ming.lei@redhat.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
(cherry picked from commit 525aa5a705d86e193726ee465d1a975265fabf19)
Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>