[netdrv] nfp: bpf: add main logics for BPF-to-BPF calls support in nfp driver
Message-id: <20190605172709.18697-4-pablo.cascon@redhat.com>
Patchwork-id: 262802
O-Subject: [RHEL8.1 BZ1713258 03/25] nfp: bpf: add main logics for BPF-to-BPF calls support in nfp driver
Bugzilla: 1713258
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Petr Oros <poros@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1713258
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=22023127

This is the main patch for the logics of BPF-to-BPF calls in the nfp
driver.

The functions called on BPF_JUMP | BPF_CALL and BPF_JUMP | BPF_EXIT were
used to call helpers and exit from the program, respectively; make them
usable for calling into, or returning from, a BPF subprogram as well.

For all calls, push the return address as well as the callee-saved
registers (R6 to R9) to the stack, and pop them upon returning from the
calls. In order to limit the overhead in terms of instruction number,
this is done through dedicated subroutines. Jumping to the callee
actually consists in jumping to the subroutine, that "returns" to the
callee: this will require some fixup for passing the address in a later
patch. Similarly, returning consists in jumping to the subroutine, which
pops registers and then return directly to the caller (but no fixup is
needed here).

Return to the caller is performed with the RTN instruction newly added
to the JIT.

For the few steps where we need to know what subprogram an instruction
belongs to, the struct nfp_insn_meta is extended with a new subprog_idx
field.

Note that checks on the available stack size, to take into account the
additional requirements associated to BPF-to-BPF calls (storing R6-R9
and return addresses), are added in a later patch.

Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>
Signed-off-by: Jiong Wang <jiong.wang@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
(cherry picked from commit 389f263b60c9a3168d47eeb44b6a07bd2b8c66e2)
Signed-off-by: Pablo Cascón <pablo.cascon@netronome.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>