[tools] selftests/powerpc: Remove Power9 copy_unaligned test
Message-id: <20190607161513.3851-3-didoming@redhat.com>
Patchwork-id: 263158
O-Subject: [RHEL8.1 PATCH 2/3] selftests/powerpc: Remove Power9 copy_unaligned test
Bugzilla: 1662374
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1662374
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=22028704

This is a test of the ISA 3.0 "copy" instruction. That instruction has
an L field, which if set to 1 specifies that "the instruction
identifies the beginning of a move group" (pp 858). That's also
referred to as "copy first" vs "copy".

In ISA 3.0B the copy instruction does not have an L field, and the
corresponding bit in the instruction must be set to 1.

This test is generating a "copy" instruction, not a "copy first", and
so on Power9 (which implements 3.0B), this results in an illegal
instruction.

So just drop the test entirely. We still have copy_first_unaligned to
test the "copy first" behaviour.

commit 83039f22ba2f6aff935a2acbb6bf671374e8317d

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Acked-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>