[include] perf/core: Declare the __percpu attribute on non-deref types
Message-id: <65d887e434c9035aad2236ad1fcbaf440f5e115f.1560250175.git.mpetlan@redhat.com>
Patchwork-id: 264906
O-Subject: [RHEL8.1 BZ1715507 003/234] perf/core: Declare the __percpu attribute on non-deref types
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit 43b9e4febc66b98d83cc1560196d56ac7fef3c32
Author: Mukesh Ojha <mojha@codeaurora.org>
Date: Tue Nov 27 14:43:32 2018 +0530

description
===========
Sparse reports the current declaration of two perf percpu variables
with this warning:

  warning: incorrect type in initializer (different address spaces)
         expected void const [noderef] <asn:3>*__vpp_verify
         got struct perf_cpu_context *<noident>

While it's normally perfectly fine to place GCC attributes anywhere
in the definition, this particular attribute is for a checking
compiler's such as Sparse's benefit, which doesn't want __percpu
on pointers.

So reorder the attribute to come after the structure type, not after
the pointer type.

[ mingo: Rewrote the changelog. ]

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>