[tools] perf trace: See if there is a map named "filtered_pids"
Message-id: <69510f340d09739e6b133cdd78c2d3c35090076a.1560250175.git.mpetlan@redhat.com>
Patchwork-id: 264950
O-Subject: [RHEL8.1 BZ1715507 020/234] perf trace: See if there is a map named "filtered_pids"
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit 744fafc787deebf26a39e6b5be36026bb2ff65fb
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Nov 7 10:04:51 2018 -0300

description
===========
Lookup for the first map named "filtered_pids" and, if augmenting
syscalls, i.e. if a BPF event is present and the
"__augmented_syscalls__" is present, then fill in that map with the pids
to filter, be it feedback loop ones (perf trace's pid, its father if it
is "sshd", more auto-filtered in the future) or the ones explicitely
stated in the tool command line via --filter-pids.

The code to actually fill in the map comes next.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>