[tools] perf probe: Fix unchecked usage of strncpy()
Message-id: <1730a53fb198976914f4b0211adb3df8e67d33f8.1560250176.git.mpetlan@redhat.com>
Patchwork-id: 264986
O-Subject: [RHEL8.1 BZ1715507 103/234] perf probe: Fix unchecked usage of strncpy()
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit bef0b8970f27da5ca223e522a174d03e2587761d
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Dec 6 11:50:08 2018 -0300

description
===========
The strncpy() function may leave the destination string buffer
unterminated, better use strlcpy() that we have a __weak fallback
implementation for systems without it.

In this case the 'target' buffer is coming from a list of build-ids that
are expected to have a len of at most (SBUILD_ID_SIZE - 1) chars, so
probably we're safe, but since we're using strncpy() here, use strlcpy()
instead to provide the intended safety checking without the using the
problematic strncpy() function.

This fixes this warning on an Alpine Linux Edge system with gcc 8.2:

  util/probe-file.c: In function 'probe_cache__open.isra.5':
  util/probe-file.c:427:3: error: 'strncpy' specified bound 41 equals destination size [-Werror=stringop-truncation]
     strncpy(sbuildid, target, SBUILD_ID_SIZE);
     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  cc1: all warnings being treated as errors

Fixes: 1f3736c9c833 ("perf probe: Show all cached probes")
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>