[tools] perf trace: Allow selecting use the use of the ordered_events code
Message-id: <5166f0dea4505c77bd5ef40f5f36c7beeb4a1904.1560250176.git.mpetlan@redhat.com>
Patchwork-id: 265003
O-Subject: [RHEL8.1 BZ1715507 118/234] perf trace: Allow selecting use the use of the ordered_events code
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit a761a8d102e2a6edfef3ad5bb22baeb715f93abd
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Dec 10 16:27:33 2018 -0300

description
===========
I was trigger happy on this one, as using ordered_events as implemented
by Jiri for use with the --block code under discussion on lkml incurs
in delaying processing to form batches that then get ordered and then
printed.

With 'perf trace' we want to process the events as they go, without that
delay, and doing it that way works well for the common case which is to
trace a thread or a workload started by 'perf trace'.

So revert back to not using ordered_events but add an option to select
that mode so that users can experiment with their particular use case to
see if works better, i.e. if the added delay is not a problem and the
ordering helps.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>