[tools] perf cs-etm: Treat EO_TRACE element as trace discontinuity
Message-id: <f65df900927a4112d7ad12ca09f76843c793a224.1560250176.git.mpetlan@redhat.com>
Patchwork-id: 264993
O-Subject: [RHEL8.1 BZ1715507 141/234] perf cs-etm: Treat EO_TRACE element as trace discontinuity
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit 02e7e2509e939b50f4f1b14c08b97f60256df3ea
Author: Leo Yan <leo.yan@linaro.org>
Date: Tue Dec 11 15:38:27 2018 +0800

description
===========
If the decoder outputs an EO_TRACE element, it means the end of the
trace buffer; this is a discontinuity and in this case the end of trace
data needs to be saved.

This patch generates a CS_ETM_DISCONTINUITY packet for the EO_TRACE
element hereby flushing the end of trace data in cs-etm.c.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>