[tools] perf trace: Show NULL when syscall pointer args are 0
Message-id: <c3b971ffa9e5cf4f1974a40d763b1bba7a6ad210.1560250176.git.mpetlan@redhat.com>
Patchwork-id: 265055
O-Subject: [RHEL8.1 BZ1715507 159/234] perf trace: Show NULL when syscall pointer args are 0
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit ce05539f20b337221d3310baf73b45317dab067c
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Dec 17 12:21:09 2018 -0300

description
===========
Matching strace's output format. The 'format' file for the syscall
tracepoints have an indication if the arg is a pointer, with some
exceptions like 'mmap' that has its first arg as an 'unsigned long', so
use a heuristic using the argument name, i.e. if it contains the 'addr'
substring, format it with the pointer formatter.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>