[tools] perf thread-stack: Fix thread stack processing for the idle task
Message-id: <62d9c11e9470dd84e2f28e3ca9bca2748b05ad61.1560250176.git.mpetlan@redhat.com>
Patchwork-id: 265096
O-Subject: [RHEL8.1 BZ1715507 194/234] perf thread-stack: Fix thread stack processing for the idle task
Bugzilla: 1715507
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1715507
https://bugzilla.redhat.com/show_bug.cgi?id=1715507

upstream
========
commit 256d92bc93fd40411a02be5cdba74a7bf91e6e09
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Dec 21 14:06:19 2018 +0200

description
===========
perf creates a single 'struct thread' to represent the idle task. That
is because threads are identified by PID and TID, and the idle task
always has PID == TID == 0.

However, there are actually separate idle tasks for each CPU. That
creates a problem for thread stack processing which assumes that each
thread has a single stack, not one stack per CPU.

Fix that by passing through the CPU number, and in the case of the idle
"thread", pick the thread stack from an array based on the CPU number.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>